+++ contracts/modules/andromeda-rates/Cargo.toml
@@ -24,7 +24,7 @@ cw20 = { workspace = true }
 cw2 = { workspace = true }
 semver = { workspace = true }
 
-andromeda-std = { workspace = true, features = ["rates"] }
+andromeda-std = { workspace = true }
 andromeda-modules = { workspace = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
