+++ contracts/modules/andromeda-rates/src/mock.rs
@@ -2,7 +2,7 @@
 
 use crate::contract::{execute, instantiate, query};
 use andromeda_modules::rates::InstantiateMsg;
-use andromeda_std::ado_contract::rates::Rate;
+use andromeda_std::ado_base::rates::Rate;
 use cosmwasm_std::Empty;
 use cw_multi_test::{Contract, ContractWrapper};
