+++ contracts/modules/andromeda-rates/src/state.rs
@@ -1,4 +1,4 @@
-use andromeda_std::ado_contract::rates::Rate;
+use andromeda_std::ado_base::rates::Rate;
 use cw_storage_plus::Map;
 
 pub const RATES: Map<&str, Rate> = Map::new("rates");
