+++ packages/andromeda-modules/Cargo.toml
@@ -19,5 +19,5 @@ cw-utils = { workspace = true }
 cw721 = { workspace = true }
 cw721-base = { workspace = true }
 
-andromeda-std = { workspace = true, features=["rates"] }
+andromeda-std = { workspace = true }
