+++ packages/std/src/ado_contract/query.rs
@@ -55,6 +55,8 @@ impl<'a> ADOContract<'a> {
                 AndromedaQuery::PermissionedActions {} => {
                     encode_binary(&self.query_permissioned_actions(deps)?)
                 }
+                #[cfg(feature = "rates")]
+                AndromedaQuery::GetRate { action } => encode_binary(&self.get_rates(deps, action)?),
                 _ => Err(ContractError::UnsupportedOperation {}),
             },
             Err(_) => Err(ContractError::UnsupportedOperation {}),
